10: hold shift to check multiple checkboxes
We selected all checkboxes and added an event listener to listen to the click,
and a function to handle the click.

We keep track of the last clicked checkbox and current clicked checkbox, in order
to loop trough them and turn inbetween on or off.

So we start the loop, is this the current or last checked if so turn inbetween on,
we loop trough all of them checking that, when it finds the other one that was clicked
it turns inbetween off again. When inbetween is on, we mark the checkbox as checked.

I added lastChecked to the conditional to make sure it existed, now when you click
shift without any box selected it doesn't select all. A lastChecked must exist in
order to select all.

I also tried to add unchecking multiple, but it leaves the former lastChecked
still checked. I need to think about it a bit more.