11: custom video player - wes
Right at the start, instead of selecting the ranges and skips separatelly
Wes selects them together (querySelectorAll).

Play: I had only hooked the play to the button, Wes also hooked it up to
the video. When updating the text of the button, he built a separate function
that listens to the video, that avoids problems caused by plugins people
may use to pause play videos. We can use .this in the function, besause
it's bound to the video.

Skip: Wes selects all the skipButtons, and when listening uses a for each
to listen to each button. Because we used 'data-skip', we can access the
dataset.skip. So we don't need to hardcode any values, and we don't need to
specify which skip button we want. Dataset solves that for us.

Range: Similar as skip, we listen to all the ranges and listen for a change
on each. We also listen to mousemose (we can set a flag, like we will do
for the progress bar, and only listen when the mouse is down). We have access
to this.name and this.value, e.g. when we click the volume it will return
this.name: volume and this.value: whatever the current value is. Like with
dataset this helps us so that we don't have to specify which button we want,
it returns those values all by itself.

Progress bar: Instead of using setInterval, Wes listens to the video
timeupdate, you can also listen to the video progress, both trigger when
the video is updating.

Progress bar scrub: We listen to a click, from there we can get the offsetX,
which tells you where you clicked on the bar. Doing a bit of math we can
update the size of the bar and video progress.
  Scrub with mouse down: We created a flag to only run the event listener
when the mouse is down. If mousedown is true, it moves on to run the scrub.