12: key sequence Detection
detecting and sanving our pressed keys is fairly straight foward, we listen
to a keyup and save the pressed keys in an array.

What is most interesting here is getting the array to be the same length
as our secret code. we use splice to do that.

Splice can add or remove items, the syntax:
splice(start, deleteCount, item1, item2, itemN)
We want to remove from the beginning as we enter new keys. For that we use:
– start (a negative one):
If negative, it will begin that many elements from the end of the array.
In our case: -secretCode.length - 1 = -5
– deleteCount:
The difference between our pressed length and our secret code length