15: localStorage first version
The beginner javascript course had an exercise like this, it's pretty
much the same.

I changed the const items to an let, to be able to change the list and
show the new list when an item is deleted.

The first thing to do was create the function to populate our items array
with the items objects. So we get the value from the form, from there we
create an object, then we push that into the items array, then we tell
everyone our array has been updated.

After that I wrote functions to store and get our items from localStorage.
When we get the items from the localStorage we let everyone know the list
has been updated.