18: adding up times with reduce (my attempt)
I think this one is pretty self-explanatory, I got the lis, then the
value inside the data-set from each.

I split the minutes from the seconds, converted them to ints, and the
minutes to seconds. Then added those, so I would have the total of
seconds per video.

After that I used reduce to get the total sum of seconds. And used new
Data to get that formatted as hours:minutes:seconds.