speech recognition
I tried doing a version first, but couldn't add the new
paragraph as the previous one ended. Wes solved that by
creating a new paragraph element if the transcript was
final.

In order for the speech recognition to start over and get
the new words, we also added a listener that listens to
the end of recognition, and then restarts it.

Since results is a list, he turned it into an array,
so we could map through it. The transcript value we
wanted is nested in results.