fixed the debounce
I wasn't using the debounce correctly:
I need to keep the debounced function inside a variable,
otherwise I would be redefining it all the time.
Debounce can only track the time if the instance is
always the same.

I'm using debounce, so the code waits a bit after
the user said "horas (hours)" to display the current
time. Since it takes a bit to form the whole sentence,
if the time is displayed right after the user said
"horas", it looks all glitchy.

I also made it so, that it stays on the page, and
the next sentence is added after it. BUT, the code
is repetitive. It should be refactored.