22: Follow along link
We made a highlight to our links, that follows the mouseenter
over the respective link.

We used this.getBoundingClientRect() to get the position and
size info. The animation comes from the CSS. We had to add
the window.scrollX and scrollY so that if the user has
scrolled, that scroll size/position is added to our highlight.
Otherwise it doesn't get the correct placement.