23: speech Synthesis
Line 66 gets the voices and populates our voices array,
through populateVoices(). '.this' in here is speechSynthesis.
We map trough the voices to build the html, join each option,
and add to the dropdown html.

To set the voices just trying to set through the name isn't
enough. We need to find the actual voice, so we use the name
to find it in our voices array.

We create toggle to control when the speechSynthesis starts and
stops. In the stop button we need to pass it in a function, our
it would run on page load.