24: sitcky nav
This one was super simple:
We get the nav.offsetTop and the window.scrollY to know
when the nav hit the top. When it does we add a fixed-nav
class to the body. Boom!

Why the body and not the nav?
You could add the fixed class to the nav, but adding it
to the body means we can target all of it's children.
So here we also add it to the site-wrap, to give it a
nice effect when transitioning.

Important:
When you add fixed to the nav it no longer takes up space
in the document, it sorts of 'float'. This is a CSS thing.
So we need to offset this amount of space we lost by
adding some padding to the body. This way we don't get a
janky transition, with the text jumping up.